Fix search behavior a bit
