+++ src/routes/index.svelte
@@ -2,7 +2,7 @@
   import IndexView from "../components/IndexView.svelte";
   import ActivityHeatmap from "../components/ActivityHeatmap.svelte";
   import PriceQuantityCharts from "../components/PriceQuantityCharts.svelte";
-  import SearchItemIndex from "../components/PriceSummary/SearchItemIndex.svelte";
+  import SearchItemIndex from "../components/PriceSummary/View.svelte";
   import FrontMatter from "../docs/FrontMatter.svx";
   import IndexDescription from "../docs/IndexDescription.svx";
   import References from "../docs/References.svx";
